+++ esmecata/__main_gseapy__.py
@@ -64,9 +64,36 @@ def main():
         help='Output directory path.',
         metavar='OUPUT_DIR')
 
+    parent_parser_grouping = argparse.ArgumentParser(add_help=False)
+    parent_parser_grouping.add_argument(
+        '--grouping',
+        dest='grouping',
+        required=True,
+        help='Grouping factor used for enrichment analysis (either "tax_rank" or "selected").',
+        metavar='STR',
+        default='tax_rank')
+
+    parent_parser_t = argparse.ArgumentParser(add_help=False)
+    parent_parser_t.add_argument(
+        '--taxon-rank',
+        dest='taxon_rank',
+        required=False,
+        help='Taxon rank to cluster observation names of EsMeCaTa together (default "phylum").',
+        metavar='INPUT_TAXON',
+        default='phylum')
+
+    parent_parser_taxa_list = argparse.ArgumentParser(add_help=False)
+    parent_parser_taxa_list.add_argument(
+        '--taxa-list',
+        dest='taxa_list',
+        required=False,
+        help='When using value "selected" for option "grouping", you have to give this file indicating the different groups of taxon to compare.',
+        metavar='STR',
+        default='tax_rank')
+
+
     parent_parser_e = argparse.ArgumentParser(add_help=False)
     parent_parser_e.add_argument(
-        '-e',
         '--enzyme',
         dest='enzyme_file',
         required=False,
@@ -76,23 +103,12 @@ def main():
 
     parent_parser_g = argparse.ArgumentParser(add_help=False)
     parent_parser_g.add_argument(
-        '-g',
         '--go',
         dest='go_file',
         required=False,
         help='Pathname to go-basic.obo file. If no path is given, it will be downloaded and put in the output folder.',
         metavar='INPUT_FILE',
         default=None)
-
-    parent_parser_t = argparse.ArgumentParser(add_help=False)
-    parent_parser_t.add_argument(
-        '-t',
-        '--taxon',
-        dest='taxon_rank',
-        required=False,
-        help='Taxon rank to cluster observation names of EsMeCaTa together (default "phylum").',
-        metavar='INPUT_TAXON',
-        default='phylum')
 
     # subparsers
     subparsers = parser.add_subparsers(
@@ -101,10 +117,11 @@ def main():
         dest='cmd')
 
     gseapy_taxon_parser = subparsers.add_parser(
-        'gseapy_taxon',
-        help='Extract enriched functions from taxon using gseapy and orsum.',
+        'gseapy_enrichr',
+        help='Extract enriched functions from groups (either chosen from tax_rank or manually selected) using gseapy enrichr and orsum.',
         parents=[
-            parent_parser_f, parent_parser_o, parent_parser_e, parent_parser_g, parent_parser_t
+            parent_parser_f, parent_parser_o, parent_parser_grouping, parent_parser_t, parent_parser_taxa_list,
+            parent_parser_e, parent_parser_g
             ],
         allow_abbrev=False)
 
@@ -131,8 +148,8 @@ def main():
     logger.addHandler(console_handler)
 
     if args.cmd == 'gseapy_taxon':
-        taxon_rank_annotation_enrichment(args.input_folder, args.output, enzyme_data_file=args.enzyme_file,
-                                         go_basic_obo_file=args.go_file, taxon_rank=args.taxon_rank)
+        taxon_rank_annotation_enrichment(args.input_folder, args.output, args.grouping, taxon_rank=args.taxon_rank, taxa_lists_file=args.taxa_list,
+                                         enzyme_data_file=args.enzyme_file, go_basic_obo_file=args.go_file)
 
     logger.info("--- Total runtime %.2f seconds ---" % (time.time() - start_time))
     logger.warning(f'--- Logs written in {log_file_path} ---')